pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "personacore2"
version = "2.0.0"
description = "Ollama-powered AI video generation suite with a premium cinematic GUI"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "PersonaCore Team" }]

dependencies = [
    "PyQt6>=6.6.0",
    "PyQt6-Qt6>=6.6.0",
    "ollama>=0.2.0",
    "httpx>=0.27.0",
    "platformdirs>=4.2.0",
    "tomli>=2.0.0; python_version < '3.11'",
    "tomli-w>=1.0.0",
    "Pillow>=10.3.0",
    "opencv-python>=4.9.0",
    "numpy>=1.26.0",
    "psutil>=5.9.0",
    "rich>=13.7.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
diffusers = [
    "torch>=2.2.0",
    "diffusers>=0.27.0",
    "transformers>=4.40.0",
    "accelerate>=0.29.0",
    "safetensors>=0.4.3",
    "xformers>=0.0.25; sys_platform != 'darwin'",
]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pytest>=8.2.0",
    "pytest-qt>=4.4.0",
]

[project.scripts]
personacore2 = "personacore.main:main"

[tool.hatch.build.targets.wheel]
packages = ["personacore"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
